+++ CHANGELOG.md
@@ -63,7 +63,7 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 * Error handler for Turnstile to handle/ignore client side errors
 * Construction partner CDE (Common Data Environment) distribution options
 * `supersede-record` development task for marking a record as superseding another record
-* Records presets for alias identifiers and constraints
+* Record presets for alias identifiers and constraints
 * Record distribution option presets and utilities, upstreamed from ODS automated records
 
 ### Fixed